api/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch==1.13.1+cpu
torchvision==0.14.1+cpu
numpy<2.0